New UART class, infomem class and new pin distribution

+++ hardware/msp430/libraries/panStamp/repeater.cpp
@@ -56,7 +56,7 @@ REPEATER::REPEATER(void)
 void REPEATER::packetHandler(SWPACKET *packet)
 {
   static bool repeatPacket = true;
-  static unsigned long currentTime;
+  static uint32_t currentTime;
 
   if (enabled)
   {
